feat(debtors) modify CS create in backend - Tweak form so that it properly updates on hour change (for some reason RHF field on change did not trigger the change in its internal state - Add some defaults to the payment form to improve UX - Extract some logic into helper file for class session route - Add remove from array method in utils
